requirements.txt
asgiref==3.2.10
confusable-homoglyphs==3.2.0
Django==3.1.2
django-bootstrap3==14.2.0
django-cors-headers==3.5.0
django-registration==2.4.1
django-tinymce==3.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
importlib-metadata==1.7.0
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
zipp==3.3.2
